DEVOPS-master/requirements.txt
Flask
flask-pymongo
python-dotenv
Werkzeug
dnspython
pytest 
